Expected.WithPropertiesFrom   & Expected.WithSelectedProperties  cleaned up
