+++ src/ExpectedObjects.Specs/TestTypes/TestDto.cs
@@ -1,0 +1,9 @@
+namespace ExpectedObjects.Specs.TestTypes
+{
+    public class TestDto
+    {
+        public string StringProperty { get; set; }
+        public TypeWithString TypeWithString { get; set; }
+        public TypeWithIEnumerable TypeWithIEnumerable { get; set; }
+    }
+}
