+++ src/ExpectedObjects.Specs/TestTypes/TestDto2.cs
@@ -1,0 +1,13 @@
+namespace ExpectedObjects.Specs.TestTypes
+{
+    public class TestDto2
+    {
+        public TestDto2()
+        {
+        }
+
+        public string Name1 { get; set; }
+        public TypeWithString Name2 { get; set; }
+        public TypeWithIEnumerable Name3 { get; set; }
+    }
+}
